+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Add new getCartoColorStylePropsForItem function [#614](https://github.com/CartoDB/carto-react/pull/614)
+
 ## 2.0
 
 ### 2.0.0-beta.3 (2023-03-06)